// src/app/features/authentication/models/icontactaddress.ts
export interface IContactAddress {
    contactId: string;
    locationId: string;
    houseNo: string;
    houseType: number;
    floorNo: number;
    apartmentNo: string;
    cityId: string;
    districtId: string;
    latitude: string;
    longitude: string;
    addressNotes: string;
    type: number;
}